+++ DDEyC-Assistant/Models/DTOs/MessageDTO.cs
@@ -1,0 +1,11 @@
+ 
+ namespace DDEyC_Assistant.Models.DTOs{
+public class MessageDto
+    {
+        public string Content { get; set; }
+        public string Role { get; set; }
+        public DateTime Timestamp { get; set; }
+    }
+
+ }
+
