+++ DDEyC-Assistant/Models/DTOs/ThreadDto.cs
@@ -1,0 +1,7 @@
+public class UserThreadDto
+    {
+        public int Id { get; set; }
+        public string ThreadId { get; set; }
+        public DateTime LastUsed { get; set; }
+        public bool IsActive { get; set; }
+    }
